midas-mcp-server/requirements-midas.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0,<5.0.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
